Store the state as a SQLite database rather than a JSON file
Using SQLite allows more efficient reading and writing.  It also
allows multiple deployments to be stored in a single state file (so we
might consider using something like ~/.charon/state.charon for all
deployments unless overriden).

The use of properties simplifies backend code a lot.  For instance,

  vm_id = charon.util.attr_property("vmId", None)

creates a backend property such that reading vm_id looks up the ‘vmId’
machine attribute in the database, while assigning to vm_id updates
the database.  So no more need to explicitly serialise/deserialise
internal state.

TODO:

- Update the EC2 backend.
- Don't use exclusive locking in operations like ‘charon deploy’.
- Add command line flags to support multiple deployments in a single
  state file.
- Re-enable parallelism: SQLite database handles cannot be used from
  different threads concurrently, so we either need to put a lock
  around the SQLite calls or use different handles per thread.
- Allow wrapping multiple updates (like property assignments) in a
  single transaction.

+++ charon/backends/virtualbox.py
@@ -32,49 +32,21 @@ class VirtualBoxState(MachineState):
     def get_type(cls):
         return "virtualbox"
 
-    def __init__(self, depl, name, log_file=sys.stderr):
-        MachineState.__init__(self, depl, name, log_file)
-        self._vm_id = None
-        self._ipv4 = None
-        self._disk = None
+    state = charon.util.attr_property("state", MachineState.MISSING, int) # override
+    private_ipv4 = charon.util.attr_property("privateIpv4", None)
+    disk = charon.util.attr_property("virtualbox.disk", None)
+    disk_attached = charon.util.attr_property("virtualbox.diskAttached", False, bool)
+    _client_private_key = charon.util.attr_property("virtualbox.clientPrivateKey", None)
+    _client_public_key = charon.util.attr_property("virtualbox.clientPublicKey", None)
+    _headless = charon.util.attr_property("virtualbox.headless", False, bool)
+
+    def __init__(self, depl, name, id, log_file=sys.stderr):
+        MachineState.__init__(self, depl, name, id, log_file)
         self._disk_attached = False
-        self._started = False
-        self._headless = False
-        self._client_private_key = None
-        self._client_public_key = None
-
-    def serialise(self):
-        x = MachineState.serialise(self)
-        if self._vm_id: x['vmId'] = self._vm_id
-        if self._ipv4: x['privateIpv4'] = self._ipv4
-
-        y = {}
-        if self._disk: y['disk'] = self._disk
-        if self._client_private_key: y['clientPrivateKey'] = self._client_private_key
-        if self._client_public_key: y['clientPublicKey'] = self._client_public_key
-        y['diskAttached'] = self._disk_attached
-        y['started'] = self._started
-        y['headless'] = self._headless
-        x['virtualbox'] = y
-
-        return x
-
-    def deserialise(self, x):
-        MachineState.deserialise(self, x)
-        self._vm_id = x.get('vmId', None)
-        self._ipv4 = x.get('privateIpv4', None)
-
-        y = x.get('virtualbox')
-        self._disk = y.get('disk', None)
-        self._disk_attached = y.get('diskAttached', False)
-        self._client_private_key = y.get('clientPrivateKey', None)
-        self._client_public_key = y.get('clientPublicKey', None)
-        self._started = y.get('started', False)
-        self._headless = y.get('headless', False)
 
     def get_ssh_name(self):
-        assert self._ipv4
-        return self._ipv4
+        assert self.private_ipv4
+        return self.private_ipv4
 
     def get_ssh_flags(self):
         key_file = "{0}/id_charon-{1}".format(self.depl.tempdir, self.name)
@@ -86,25 +58,17 @@ class VirtualBoxState(MachineState):
     def get_physical_spec(self, machines):
         return ['    require = [ <charon/virtualbox-image-charon.nix> ];']
 
-    @property
-    def vm_id(self):
-        return self._vm_id
-
-    @property
-    def private_ipv4(self):
-        return self._ipv4
-
 
     def address_to(self, m):
         if isinstance(m, VirtualBoxState):
-            return m._ipv4
+            return m.private_ipv4
         return MachineState.address_to(self, m)
 
 
     def _get_vm_info(self):
         '''Return the output of ‘VBoxManage showvminfo’ in a dictionary.'''
         lines = self._logged_exec(
-            ["VBoxManage", "showvminfo", "--machinereadable", self._vm_id],
+            ["VBoxManage", "showvminfo", "--machinereadable", self.vm_id],
             capture_stdout=True, check=False).splitlines()
         # We ignore the exit code, because it may be 1 while the VM is
         # shutting down (even though the necessary info is returned on
@@ -128,54 +92,50 @@ class VirtualBoxState(MachineState):
 
     def _start(self):
         self._logged_exec(
-            ["VBoxManage", "guestproperty", "set", self._vm_id, "/VirtualBox/GuestInfo/Net/1/V4/IP", ''])
+            ["VBoxManage", "guestproperty", "set", self.vm_id, "/VirtualBox/GuestInfo/Net/1/V4/IP", ''])
 
         self._logged_exec(
-            ["VBoxManage", "guestproperty", "set", self._vm_id, "/VirtualBox/GuestInfo/Charon/ClientPublicKey", self._client_public_key])
+            ["VBoxManage", "guestproperty", "set", self.vm_id, "/VirtualBox/GuestInfo/Charon/ClientPublicKey", self._client_public_key])
 
-        self._logged_exec(["VBoxManage", "startvm", self._vm_id] +
+        self._logged_exec(["VBoxManage", "startvm", self.vm_id] +
                           (["--type", "headless"] if self._headless else []))
 
-        self._started = True
-        self._state = self.STARTING
-        self.write()
+        self.state = self.STARTING
+
+
+    def _update_ip(self):
+        res = self._logged_exec(
+            ["VBoxManage", "guestproperty", "get", self.vm_id, "/VirtualBox/GuestInfo/Net/1/V4/IP"],
+            capture_stdout=True).rstrip()
+        if res[0:7] != "Value: ": return
+        self.private_ipv4 = res[7:]
 
 
     def _wait_for_ip(self):
         self.log_start("waiting for IP address...")
         while True:
-            res = self._logged_exec(
-                ["VBoxManage", "guestproperty", "get", self._vm_id, "/VirtualBox/GuestInfo/Net/1/V4/IP"],
-                capture_stdout=True).rstrip()
-            if res[0:7] == "Value: ":
-                self._ipv4 = res[7:]
-                self.log_end(" " + self._ipv4)
-                break
+            self._update_ip()
+            if self.private_ipv4 != None: break
             time.sleep(1)
             self.log_continue(".")
-
-        charon.known_hosts.remove(self._ipv4)
-
-        self.write()
+        self.log_end(" " + self.private_ipv4)
+        charon.known_hosts.remove(self.private_ipv4)
 
 
     def create(self, defn, check, allow_reboot):
         assert isinstance(defn, VirtualBoxDefinition)
 
-        if self._state != self.UP: check = True
+        if self.state != self.UP or check: self.check()
 
-        if not self._vm_id:
+        if not self.vm_id:
             self.log("creating VirtualBox VM...")
-
             vm_id = "charon-{0}-{1}".format(self.depl.uuid, self.name)
-
             self._logged_exec(["VBoxManage", "createvm", "--name", vm_id, "--ostype", "Linux", "--register"])
+            self.vm_id = vm_id
+            self.state = self.STOPPED
 
-            self._vm_id = vm_id
-            self.write()
-
-        if not self._disk:
-            vm_dir = os.environ['HOME'] + "/VirtualBox VMs/" + self._vm_id
+        if not self.disk:
+            vm_dir = os.environ['HOME'] + "/VirtualBox VMs/" + self.vm_id
             if not os.path.isdir(vm_dir):
                 raise Exception("can't find directory of VirtualBox VM ‘{0}’".format(self.name))
 
@@ -196,37 +156,27 @@ class VirtualBoxState(MachineState):
 
             self._logged_exec(["VBoxManage", "clonehd", base_image, disk])
 
-            self._disk = disk
-            self.write()
+            self.disk = disk
 
-        if not self._disk_attached:
+        if not self.disk_attached:
             self._logged_exec(
-                ["VBoxManage", "storagectl", self._vm_id,
+                ["VBoxManage", "storagectl", self.vm_id,
                  "--name", "SATA", "--add", "sata", "--sataportcount", "2",
                  "--bootable", "on", "--hostiocache", "on"])
 
             self._logged_exec(
-                ["VBoxManage", "storageattach", self._vm_id,
+                ["VBoxManage", "storageattach", self.vm_id,
                  "--storagectl", "SATA", "--port", "0", "--device", "0",
-                 "--type", "hdd", "--medium", self._disk])
+                 "--type", "hdd", "--medium", self.disk])
 
-            self._disk_attached = True
-            self.write()
-
-        if check:
-            if self._get_vm_state() == 'running':
-                self._started = True
-            else:
-                self.log("VirtualBox VM went down, restarting...")
-                self._started = False
-                self.write()
+            self.disk_attached = True
 
         if not self._client_private_key:
             (self._client_private_key, self._client_public_key) = self._create_key_pair()
 
-        if not self._started:
+        if not self.started:
             self._logged_exec(
-                ["VBoxManage", "modifyvm", self._vm_id,
+                ["VBoxManage", "modifyvm", self.vm_id,
                  "--memory", defn.memory_size, "--vram", "10",
                  "--nictype1", "virtio", "--nictype2", "virtio",
                  "--nic2", "hostonly", "--hostonlyadapter2", "vboxnet0",
@@ -235,24 +185,28 @@ class VirtualBoxState(MachineState):
             self._headless = defn.headless
             self._start()
 
-        if not self._ipv4 or check:
+        if not self.private_ipv4 or check:
             self._wait_for_ip()
 
 
     def destroy(self):
+        if not self.vm_id: return True
+
         if not self.depl.confirm("are you sure you want to destroy VirtualBox VM ‘{0}’?".format(self.name)): return False
 
         self.log("destroying VirtualBox VM...")
 
         if self._get_vm_state() == 'running':
-            self._logged_exec(["VBoxManage", "controlvm", self._vm_id, "poweroff"], check=False)
+            self._logged_exec(["VBoxManage", "controlvm", self.vm_id, "poweroff"], check=False)
 
         while self._get_vm_state() not in ['poweroff', 'aborted']:
             time.sleep(1)
 
+        self.state = self.STOPPED
+
         time.sleep(1) # hack to work around "machine locked" errors
 
-        self._logged_exec(["VBoxManage", "unregistervm", "--delete", self._vm_id])
+        self._logged_exec(["VBoxManage", "unregistervm", "--delete", self.vm_id])
 
         return True
 
@@ -263,8 +217,7 @@ class VirtualBoxState(MachineState):
         self.log_start("shutting down... ")
 
         self.run_command("poweroff &")
-        self._state = self.STOPPING
-        self.write()
+        self.state = self.STOPPING
 
         while True:
             state = self._get_vm_state()
@@ -274,35 +227,32 @@ class VirtualBoxState(MachineState):
 
         self.log_end("")
 
-        self._started = False
-        self._state = self.STOPPED
-        self.write()
+        self.state = self.STOPPED
 
 
     def start(self):
         if self._get_vm_state() == 'running': return
         self.log("restarting...")
 
-        prev_ipv4 = self._ipv4
+        prev_ipv4 = self.private_ipv4
 
         self._start()
         self._wait_for_ip()
 
-        if prev_ipv4 != self._ipv4:
+        if prev_ipv4 != self.private_ipv4:
             self.warn("IP address has changed, you may need to run ‘charon deploy’")
 
         self.wait_for_ssh(check=True)
 
 
     def check(self):
-        old_state = self._state
+        if not self.vm_id: return
         state = self._get_vm_state()
         self.log("VM state is ‘{0}’".format(state))
         if state == "poweroff" or state == "aborted":
-            self._state = self.STOPPED
+            self.state = self.STOPPED
         elif state == "running":
+            self._update_ip()
             MachineState.check(self)
         else:
-            self._state = self.UNKNOWN
-        if old_state != self._state:
-            self.write()
+            self.state = self.UNKNOWN
